OnSizeAllocated and bounds positioning moved back to VisualElementRenderer
