+++ Xamarin.Forms.Platform.GTK/Renderers/LayoutRenderer.cs
@@ -1,5 +1,4 @@
 ﻿using Gtk;
-using Xamarin.Forms.Platform.GTK.Extensions;
 using Xamarin.Forms.Platform.GTK.Packagers;
 
 namespace Xamarin.Forms.Platform.GTK.Renderers
@@ -8,7 +7,6 @@ namespace Xamarin.Forms.Platform.GTK.Renderers
     {
         private Fixed _fixed;
         private LayoutElementPackager _packager;
-        private Gdk.Rectangle _lastAllocation;
 
         protected override void OnElementChanged(ElementChangedEventArgs<Layout> e)
         {
@@ -33,31 +31,6 @@ namespace Xamarin.Forms.Platform.GTK.Renderers
             }
 
             base.OnElementChanged(e);
-        }
-
-        protected override void OnSizeAllocated(Gdk.Rectangle allocation)
-        {
-            base.OnSizeAllocated(allocation);
-
-            if (_lastAllocation != allocation)
-            {
-                _lastAllocation = allocation;
-
-                Rectangle bounds = Element.Bounds;
-                Container.MoveTo((int)bounds.X, (int)bounds.Y);
-
-                for (var i = 0; i < ElementController.LogicalChildren.Count; i++)
-                {
-                    var child = ElementController.LogicalChildren[i] as VisualElement;
-
-                    if (child != null)
-                    {
-                        var renderer = Platform.GetRenderer(child);
-                        renderer?.Container.SetSize(child.Bounds.Width, child.Bounds.Height);
-                        renderer?.Container.MoveTo(child.Bounds.X, child.Bounds.Y);
-                    }
-                }
-            }
         }
     }
 }
