+++ Xamarin.Forms.Platform.GTK/Renderers/ScrollViewRenderer.cs
@@ -9,7 +9,6 @@ namespace Xamarin.Forms.Platform.GTK.Renderers
     {
         private VisualElement _currentView;
         private Viewport _viewPort;
-        private Gdk.Rectangle _lastAllocation;
 
         protected IScrollViewController Controller
         {
@@ -54,31 +53,6 @@ namespace Xamarin.Forms.Platform.GTK.Renderers
                 UpdateOrientation();
                 LoadContent();
                 UpdateContentSize();
-            }
-        }
-
-        protected override void OnSizeAllocated(Gdk.Rectangle allocation)
-        {
-            base.OnSizeAllocated(allocation);
-
-            if (_lastAllocation != allocation)
-            {
-                _lastAllocation = allocation;
-
-                Rectangle bounds = Element.Bounds;
-                Container.MoveTo((int)bounds.X, (int)bounds.Y);
-
-                for (var i = 0; i < ElementController.LogicalChildren.Count; i++)
-                {
-                    var child = ElementController.LogicalChildren[i] as VisualElement;
-
-                    if (child != null)
-                    {
-                        var renderer = Platform.GetRenderer(child);
-                        renderer?.Container.SetSize(child.Bounds.Width, child.Bounds.Height);
-                        renderer?.Container.MoveTo(child.Bounds.X, child.Bounds.Y);
-                    }
-                }
             }
         }
