+++ Xamarin.Forms.Platform.GTK/VisualElementRenderer.cs
@@ -18,6 +18,7 @@ namespace Xamarin.Forms.Platform.GTK
         private VisualElementTracker<TElement, TNativeElement> _tracker;
         private string _defaultAccessibilityLabel;
         private string _defaultAccessibilityHint;
+        private Gdk.Rectangle _lastAllocation;
 
         protected VisualElementRenderer()
         {
@@ -133,6 +134,31 @@ namespace Xamarin.Forms.Platform.GTK
             base.Dispose();
 
             Dispose(true);
+        }
+
+        protected override void OnSizeAllocated(Gdk.Rectangle allocation)
+        {
+            base.OnSizeAllocated(allocation);
+
+            if (_lastAllocation != allocation)
+            {
+                _lastAllocation = allocation;
+
+                Rectangle bounds = Element.Bounds;
+                Container.MoveTo((int)bounds.X, (int)bounds.Y);
+
+                for (var i = 0; i < ElementController.LogicalChildren.Count; i++)
+                {
+                    var child = ElementController.LogicalChildren[i] as VisualElement;
+
+                    if (child != null)
+                    {
+                        var renderer = Platform.GetRenderer(child);
+                        renderer?.Container.SetSize(child.Bounds.Width, child.Bounds.Height);
+                        renderer?.Container.MoveTo(child.Bounds.X, child.Bounds.Y);
+                    }
+                }
+            }
         }
 
         protected virtual void OnRegisterEffect(PlatformEffect effect)
